% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartex_score.R
\name{cartex_score}
\alias{cartex_score}
\title{Calculate cartex score described in the paper}
\usage{
cartex_score(gene_counts, metadata = NULL)
}
\arguments{
\item{gene_counts}{A normalized gene counts data frame with genes as rows and samples as columns.}

\item{metadata}{An optional metadata dataframe corresponding to samples in the gene count dataframe}
}
\value{
A dataframe of cartex scores.
}
\description{
Calculate cartex score described in the paper
}
