% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_score.R
\name{compute_score}
\alias{compute_score}
\title{Calculate PCA based phenotype specific score}
\usage{
compute_score(
  gene_counts,
  genelist = NULL,
  num_mv_genes = NULL,
  pca_number = 1
)
}
\arguments{
\item{gene_counts}{A normalized gene counts data frame with genes as rows and samples as columns.}

\item{genelist}{An optional character vector of gene names, Default is NULL}

\item{num_mv_genes}{An optional numerical variable specifying the number of most variable genes selection , default all genes from genelist or gene_counts if genelist if null, are used to compute score}

\item{pca_number}{An integer specifying which Principal Component to use for weights. Default is 1.}
}
\value{
A named vector of scores.
}
\description{
This function computes the score based on gene variance and weights
using principal component analysis (PCA) of a gene expression matrix
}
