% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weights.R
\name{get_weights}
\alias{get_weights}
\title{Calculate Gene Weights from PCA}
\usage{
get_weights(gene_counts, genelist = NULL, pca_number = 1)
}
\arguments{
\item{gene_counts}{A normalized gene counts data frame with genes as rows and samples as columns.}

\item{genelist}{An optional character vector of gene names to subset gene counts, Default is NULL}

\item{pca_number}{An integer specifying which Principal Component to use for weights. Default is 1.}
}
\value{
gene weights based on PCA
}
\description{
This function computes weights using principal component analysis (PCA)
on the input gene expression matrix
}
