% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartex_sc.R
\name{print.cartex_report}
\alias{print.cartex_report}
\title{Print a CARTEx transparency report}
\usage{
\method{print}{cartex_report}(x, ...)
}
\arguments{
\item{x}{A \code{cartex_report} object.}

\item{...}{Ignored.}
}
\description{
Print a CARTEx transparency report
}
