% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_cartex.R
\name{single_cell_cartex}
\alias{single_cell_cartex}
\title{Apply CARTEx Score to Each Cell in a Seurat Object}
\usage{
single_cell_cartex(seurat_obj, assay = "RNA")
}
\arguments{
\item{seurat_obj}{A Seurat object containing single-cell RNA-seq data.}

\item{assay}{The assay in the Seurat object to use (default = "RNA").}
}
\value{
A Seurat object with CARTEx scores added to meta data.
}
\description{
This function computes the CARTEx score for each cell in a Seurat object
}
